Adding login, user sessions and to-dos by user
Basic course, and functionalities of Next completed

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import prisma from "@/app/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import * as yup from "yup"
+import { getUserSessionServer } from "../auth/actions/auth-actions";
 
 export async function GET(request: Request) {
 
@@ -31,10 +32,13 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+  const user = await getUserSessionServer()
+  if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })
+
   try {
     const { complete, description } = await postSchema.validate(await request.json())
 
-    const toDo = await prisma.toDo.create({ data: { complete, description } })
+    const toDo = await prisma.toDo.create({ data: { complete, description, userId: user.id } })
 
     return NextResponse.json(toDo)
   } catch (error) {
@@ -45,9 +49,12 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+  const user = await getUserSessionServer()
+  if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })
+
   try {
 
-  await prisma.toDo.deleteMany({ where: {complete: true}  })
+  await prisma.toDo.deleteMany({ where: {complete: true, userId: user.id }  })
 
     return NextResponse.json('Deleted all completed ToDos')
   } catch (error) {
